야야

+++ OpenApi/Dto/REAL10001_Data.cs
@@ -11,7 +11,7 @@ namespace OpenApi.Dto
         public string 현재시간 = "";
         public int 현재가 = 0;
         public int 전일대비 = 0;
-        public int 등락율 = 0;
+        public float 등락율 = 0;
         public int 매도호가 = 0;
         public int 매수호가 = 0;
         public int 누적거래량 = 0;
@@ -21,9 +21,9 @@ namespace OpenApi.Dto
         public int 저가 = 0;
         public int 전일대비기호 = 0;
         public int 전일거래량대비_계약_주 = 0;
-        public int 거래대금증감 = 0;
-        public int 전일거래량대비_비율 = 0;
-        public int 거래회전율 = 0;
+        public decimal 거래대금증감 = 0;
+        public float 전일거래량대비_비율 = 0;
+        public float 거래회전율 = 0;
         public int 거래비용 = 0;
         public int 시가총액_억 = 0;
         public int 상한가발생시간 = 0;
